feat(guild-race): automate race take/progress with CD and plantability gates
Parse FML race state, drive in-progress tasks into the planner, and skip pool
takes that are far-CD or unplantable plant-harvest targets.

+++ internal/automation/types.go
@@ -261,7 +261,7 @@ type PlannedOp struct {
 	BatchID        int32
 	SlotID         int32
 	TaskID         int32
-	TaskMsID       int32
+	TaskMsID       int64
 	MilestoneIndex int32
 	TargetID       int32
 	ItemID         int32
